refactor: Unify CardTabs
